digit-recognition/requirements.txt
numpy
scipy
tensorflow
tensorflow-gpu
keras
six
protobuf
graphviz
nltk
matplotlib
sklearn
scipy
pydot
graphviz
gensim
pandas
pytz
pyemd
keras_bert
ann-visualizer
python-mnist